requirements.txt
aiogram==3.4.1
gspread==5.12.0
google-generativeai==0.3.2
python-dotenv==1.0.0
apscheduler==3.10.4
aiohttp==3.9.1
pandas==2.1.4
openpyxl==3.1.2
matplotlib==3.8.2
reportlab==4.0.7
sentry-sdk==1.39.2
redis==5.0.1